% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datBRSP.gene.array}
\alias{datBRSP.gene.array}
\title{Microarray RNA Expression Values for the Human Brain Span Atlas.}
\format{A data frame with 8502744 rows and 4 variables:
\describe{
 \item{gene}{gene from the brain span atals }
 \item{brain_structure}{brain structure acronym}
 \item{value}{normalized log2 expression value}
 \item{age}{specimen age; (gene, brain_structure, age) are unique}
}}
\source{
\url{"http://www.brainspan.org/static/download.html"}
}
\usage{
datBRSP.gene.array
}
\description{
Data frame containing the normalized RNA values for the genes studied via micro array sampling in the brain span altas.
}
\keyword{datasets}
