% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genesBRSP}
\alias{genesBRSP}
\title{Genes in the brain span atlas}
\format{A vector with 48466}
\source{
\url{"http://www.brainspan.org/static/download.html"}
}
\usage{
genesBRSP
}
\description{
Vector of genes in the in the brain span atlas
}
\keyword{datasets}
